Correção exercícios 25 e 27

+++ ex25.java
@@ -7,10 +7,10 @@ public class ex25 {
 	public static void main(String[] args) {
 		Scanner ler = new Scanner(System.in);
 		
-		double a;
+		int a;
 		
 		System.out.println("Digite um valor para verificar se é par ou ímpar: ");
-		a = ler.nextDouble();
+		a = ler.nextInt();
 		
 		if ( (a % 2) == 0) {
 			System.out.println("O valor digitado (" + a + ") é par.");
